fixed type casting in i32 operations

+++ core/vm/i32Opcodes.go
@@ -1,71 +1,70 @@
 package vm
 
-type i32Sub struct {}
+type i32Sub struct{}
 
 func (op i32Sub) doOp(m *Machine) {
-	a := m.popFromStack()
-	b := m.popFromStack()
-	m.pushToStack(b - a)
+	a := uint32(uint32(m.popFromStack()))
+	b := uint32(uint32(m.popFromStack()))
+	m.pushToStack(uint64(uint32(b - a)))
 	m.pointInCode++
 }
 
-type i32Add struct {}
+type i32Add struct{}
 
 func (op i32Add) doOp(m *Machine) {
-	m.pushToStack(m.popFromStack() + m.popFromStack())
+	m.pushToStack(uint64(uint32(uint32(m.popFromStack()) + uint32(m.popFromStack()))))
 	m.pointInCode++
 }
 
-type i32Mul struct {}
+type i32Mul struct{}
 
 func (op i32Mul) doOp(m *Machine) {
-	m.pushToStack(m.popFromStack() * m.popFromStack())
+	m.pushToStack(uint64(uint32(uint32(m.popFromStack()) * uint32(m.popFromStack()))))
 	m.pointInCode++
 }
 
-type i32Xor struct {}
+type i32Xor struct{}
 
 func (op i32Xor) doOp(m *Machine) {
-	m.pushToStack(m.popFromStack() ^ m.popFromStack())
+	m.pushToStack(uint64(uint32(uint32(m.popFromStack()) ^ uint32(m.popFromStack()))))
 	m.pointInCode++
 }
 
-type i32Or struct {}
+type i32Or struct{}
 
 func (op i32Or) doOp(m *Machine) {
-	m.pushToStack(m.popFromStack() | m.popFromStack())
+	m.pushToStack(uint64(uint32(uint32(m.popFromStack()) | uint32(m.popFromStack()))))
 	m.pointInCode++
 }
 
-type i32And struct {}
+type i32And struct{}
 
 func (op i32And) doOp(m *Machine) {
-	m.pushToStack(m.popFromStack() & m.popFromStack())
+	m.pushToStack(uint64(uint32(uint32(m.popFromStack()) & uint32(m.popFromStack()))))
 	m.pointInCode++
 }
 
-type i32Remu struct {} 
+type i32Remu struct{}
 
 func (op i32Remu) doOp(m *Machine) {
-	i1 := m.popFromStack()
-	i2 := m.popFromStack()
+	i1 := uint32(m.popFromStack())
+	i2 := uint32(m.popFromStack())
 
 	if i1 != 0 {
-		m.pushToStack(i1 % i2)
+		m.pushToStack(uint64(uint32(i1 % i2)))
 	}
 
 	m.pointInCode++
 }
 
-
-type i32Divu struct {}
+type i32Divu struct{}
 
 func (op i32Divu) doOp(m *Machine) {
-	i1 := m.popFromStack()
-	i2 := m.popFromStack()
+	i1 := uint32(m.popFromStack())
+	i2 := uint32(m.popFromStack())
 
 	if i1 != 0 {
-		m.pushToStack(i1 / i2)
+		m.pushToStack(uint64(uint32(i1 / i2)))
 	}
 	m.pointInCode++
 }
@@ -75,6 +74,6 @@ type i32Const struct {
 }
 
 func (op i32Const) doOp(m *Machine) {
-	m.pushToStack(uint64(op.val))
+	m.pushToStack(uint64(uint32(op.val)))
 	m.pointInCode++
 }
